Move Extension methods to Extensions namespace use MediaTypeNames.Application.Json instead of MediaTypeHeaderValue.Parse("application/json").MediaType remove unused classes and redundant comments Attempt to implement a docker compose to smooth spinning up local dev services.
