+++ docker-compose.yml
@@ -1,0 +1,51 @@
+networks:
+  gpcac: {}
+
+services:
+
+  migrator:
+    build:
+      context: ./database
+      dockerfile: Dockerfile
+    networks:
+      - gpcac
+    depends_on:
+      - db
+    environment:
+      FLYWAY_PASSWORD: postgres
+    command:
+      - "-url=jdbc:postgresql://db/gpcac-local"
+      - "-user=postgres"
+      - "-connectRetries=10"
+      - "migrate"
+    restart: 'no'
+
+  db:
+    image: postgres:16.8-alpine
+    environment:
+      - POSTGRES_USER=postgres
+      - POSTGRES_PASSWORD=postgres
+      - POSTGRES_DB=gpcac-local
+    networks:
+      - gpcac
+    ports:
+      - '5432:5432'
+    volumes:
+      - ../DockerData/gpcac-postgres:/var/lib/postgresql/data
+
+  vk:
+    image: valkey/valkey:8.1.0
+    ports:
+      - '6379:6379'
+    networks:
+      - gpcac
+
+#__api__
+#  cd modules/api
+#  docker build -t gpconnect-appointment-checker-api .
+#  docker run -d --network gpconnect-appointment-checker_gpcac -p 8000:8080 -e "ConnectionStrings:DefaultConnection=Server=db;Port=5432;Database=pgcac-local;User Id=postgres;Password=postgres" --name gpconnect-appointment-checker-api gpconnect-appointment-checker-api
+
+#__end_user__
+#  cd modules/end-user
+#  docker build -t gpconnect-appointment-checker-end-user .
+#  docker run -d --network gpconnect-appointment-checker_gpcac -p 8001:8080 -e "ASPNETCORE_ENVIRONMENT=local" --name gpconnect-appointment-checker-end-user gpconnect-appointment-checker-end-user
