+++ modules/api/src/nlog.config
@@ -6,10 +6,16 @@
         <add assembly="NLog.Web.AspNetCore"/>
     </extensions>
     <targets>
-        <target name="console" xsi:type="Console">
+        <target name="console" xsi:type="ColoredConsole">
+            <highlight-word foregroundColor="Red" ignoreCase="true" text="error" wholeWords="true" />
+            <highlight-word foregroundColor="Magenta" ignoreCase="true" text="warn" wholeWords="true" />
+            <highlight-word foregroundColor="Blue" ignoreCase="true" text="info" wholeWords="true" />
+
             <layout xsi:type="JsonLayout" includeEventProperties="false">
                     <attribute name="time" layout="${longdate:universalTime=true}" />
+                    <attribute name="environment" layout="${configsetting:item=ASPNETCORE_ENVIRONMENT}" />
                     <attribute name="application" layout="${configsetting:item=ApplicationName}" />
+                    <attribute name="userId" layout="${aspnet-user-claim:ClaimTypes.NameIdentifier}" />
                     <attribute name="logger" layout="${logger}"/>
                     <attribute name="level" layout="${level:upperCase=true}"/>
                     <attribute name="message" layout="${message}" />
