+++ modules/end-user/src/gpconnect-appointment-checker.Helpers/CustomValidations/MaximumNumberOfCodes.cs
@@ -2,6 +2,8 @@
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 
+using gpconnect_appointment_checker.Helpers.Extensions;
+
 namespace gpconnect_appointment_checker.Helpers.CustomValidations
 {
     public class MaximumNumberOfCodesAttribute : ValidationAttribute
@@ -22,7 +24,7 @@ namespace gpconnect_appointment_checker.Helpers.CustomValidations
         {
             HttpContextAccessor = (IHttpContextAccessor)validationContext.GetService(typeof(IHttpContextAccessor));
             var multiSearchEnabled = (HttpContextAccessor?.HttpContext.User.GetClaimValue("MultiSearchEnabled").StringToBoolean(false)).GetValueOrDefault();
-            var propertyTestedInfo = validationContext.ObjectType.GetProperty(this._dependentProperty);
+            var propertyTestedInfo = validationContext.ObjectType.GetProperty(_dependentProperty);
             if (propertyTestedInfo == null)
             {
                 return new ValidationResult($"Unknown Property {_dependentProperty}");
