+++ modules/end-user/src/gpconnect-appointment-checker.Helpers/Enumerations/EnumHelper.cs
