+++ modules/end-user/src/gpconnect-appointment-checker.Helpers/Extensions/HttpClientExtensions.cs
@@ -2,7 +2,7 @@
 using System.Net.Http;
 using System.Threading.Tasks;
 
-namespace gpconnect_appointment_checker.Helpers;
+namespace gpconnect_appointment_checker.Helpers.Extensions;
 
 public static class HttpClientExtensions
 {
