+++ modules/end-user/src/gpconnect-appointment-checker.Helpers/Extensions/NumberExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace gpconnect_appointment_checker.Helpers
+namespace gpconnect_appointment_checker.Helpers.Extensions
 {
     public static class NumberExtensions
     {
