+++ modules/end-user/src/gpconnect-appointment-checker/Core/Configuration/Infrastructure/Authentication/PrincipalValidator.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Authentication.Cookies;
 using System;
 using System.Threading.Tasks;
 
+using gpconnect_appointment_checker.Helpers.Extensions;
+
 namespace gpconnect_appointment_checker.Configuration.Infrastructure.Authentication
 {
     internal static class PrincipalValidator
