--- modules/end-user/src/gpconnect-appointment-checker/Core/Configuration/Infrastructure/Logging/Interface/ILoggerManager.cs
@@ -1,10 +1,0 @@
-﻿namespace gpconnect_appointment_checker.Configuration.Infrastructure.Logging.Interface
-{
-    public interface ILoggerManager
-    {
-        void LogInformation(string message);
-        void LogWarning(string message);
-        void LogDebug(string message);
-        void LogError(string message);
-    }
-}
