--- modules/end-user/src/gpconnect-appointment-checker/Core/Configuration/Infrastructure/Logging/LoggingExtensions.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using NLog;
-
-namespace gpconnect_appointment_checker.Configuration.Infrastructure
-{
-    public static class LoggingExtensions
-    {
-        public static IServiceCollection ConfigureLoggingServices(this IServiceCollection services, IConfiguration configuration)
-        {
-            var nLogConfiguration = new NLog.Config.LoggingConfiguration();
-
-            nLogConfiguration.Variables.Add(Helpers.Constants.Headers.UserId, null);
-
-            LogManager.Configuration = nLogConfiguration;
-
-            services.AddLogging(builder =>
-            {
-                builder.ClearProviders();
-                builder.AddConfiguration(configuration.GetSection("Logging"));
-            });
-
-            return services;
-        }
-    }
-}
