+++ modules/end-user/src/gpconnect-appointment-checker/Core/HttpClientServices/ReportingService.cs
@@ -13,8 +13,11 @@ using System.Collections.Generic;
 using System.Data;
 using System.Net.Http;
 using System.Net.Http.Headers;
+using System.Net.Mime;
 using System.Text;
 using System.Threading.Tasks;
+
+using gpconnect_appointment_checker.Helpers.Extensions;
 
 namespace GpConnect.AppointmentChecker.Core.HttpClientServices;
 
@@ -52,7 +55,7 @@ public class ReportingService : IReportingService
         var json = new StringContent(
             JsonConvert.SerializeObject(reportExport, null, _options),
             Encoding.UTF8,
-            MediaTypeHeaderValue.Parse("application/json").MediaType);
+            MediaTypeNames.Application.Json);
 
         var response = await _httpClient.PostWithHeadersAsync("/reporting/export", new Dictionary<string, string>()
         {
