+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Index.cshtml
@@ -2,6 +2,7 @@
 @using gpconnect_appointment_checker.Helpers.Constants
 @using gpconnect_appointment_checker.Helpers;
 @using gpconnect_appointment_checker.Helpers.Enumerations;
+@using gpconnect_appointment_checker.Helpers.Extensions
 @model IndexModel
 @addTagHelper *, gpconnect-appointment-checker
