+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Pending/CreateAccount.cshtml.cs
@@ -1,7 +1,6 @@
 using GpConnect.AppointmentChecker.Core.Configuration;
 using GpConnect.AppointmentChecker.Core.HttpClientServices.Interfaces;
 using GpConnect.AppointmentChecker.Models.Request;
-using gpconnect_appointment_checker.Configuration.Infrastructure.Logging.Interface;
 using gpconnect_appointment_checker.Helpers;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -9,6 +8,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using System.Threading.Tasks;
 using gpconnect_appointment_checker.Core.HttpClientServices.Interfaces;
+using gpconnect_appointment_checker.Helpers.Extensions;
 
 namespace gpconnect_appointment_checker.Pages
 {
@@ -16,17 +16,15 @@ namespace gpconnect_appointment_checker.Pages
     {
         protected ILogger<CreateAccountModel> _logger;
         protected IHttpContextAccessor _contextAccessor;
-        protected readonly ILoggerManager _loggerManager;
-        private readonly IUserService _userService;        
+        private readonly IUserService _userService;
 
-        public CreateAccountModel(IOptions<GeneralConfig> configuration, IHttpContextAccessor contextAccessor, ILogger<CreateAccountModel> logger, IUserService userService, ILoggerManager loggerManager = null) : base(configuration, contextAccessor)
+        public CreateAccountModel(IOptions<GeneralConfig> configuration,
+            IHttpContextAccessor contextAccessor,
+            ILogger<CreateAccountModel> logger,
+            IUserService userService) : base(configuration, contextAccessor)
         {
             _contextAccessor = contextAccessor;
             _userService = userService;
-            if (null != loggerManager)
-            {
-                _loggerManager = loggerManager;
-            }
         }
 
         public IActionResult OnGet()
