+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Private/Logout.cshtml
@@ -3,6 +3,7 @@
 @using gpconnect_appointment_checker.Helpers.Constants
 @using gpconnect_appointment_checker.Helpers.Enumerations
 @using gpconnect_appointment_checker.Helpers
+@using gpconnect_appointment_checker.Helpers.Extensions
 @{
     var userAccountStatus = User.GetClaimValue<UserAccountStatus>("UserAccountStatus");
 }
