+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Private/SearchDetailResults.cshtml
@@ -1,5 +1,6 @@
 ﻿@using gpconnect_appointment_checker.Helpers
 @using gpconnect_appointment_checker.Helpers.Constants
+@using gpconnect_appointment_checker.Helpers.Extensions
 @model SearchDetailModel
 
 <div class="nhsuk-grid-row">
