+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Private/StatsCount.cshtml
@@ -1,5 +1,6 @@
 ﻿@using gpconnect_appointment_checker.Helpers.Constants
 @using gpconnect_appointment_checker.Helpers
+@using gpconnect_appointment_checker.Helpers.Extensions
 
 @if (Model.SearchResultsTotalCount > 0)
 {
