+++ modules/end-user/src/gpconnect-appointment-checker/Pages/Public/Navigation.cshtml
@@ -1,6 +1,7 @@
 ﻿@using gpconnect_appointment_checker.Helpers.Constants
 @using gpconnect_appointment_checker.Helpers.Enumerations
 @using gpconnect_appointment_checker.Helpers
+@using gpconnect_appointment_checker.Helpers.Extensions
 @{
     var userIsAdmin = User.GetClaimValue("IsAdmin").StringToBoolean(false);
     var userAccountStatus = User.GetClaimValue<UserAccountStatus>("UserAccountStatus");
